Function to calculate the error in estimated sample size for use in optimize function

Function to calculate the error in estimated sample size for use in optimize function

relR_samplesize_opterr(
  m,
  R_a,
  R_b,
  p_a,
  N,
  alpha,
  alternative,
  power,
  sensitivity,
  specificity,
  overdispersion
)

Arguments

m

the sample size.

R_a

Numeric (Positive). The assumed R among the group in the denominator of the ratio. Input value must be greater than 0.

R_b

Numeric (Positive). The assumed R among the group in the numerator of the ratio. Input value must be greater than 0.

p_a

Numeric. The proportion of the population in group a. Must be between 0 and 1.

N

Numeric (Positive). The size of the infected pool. Only one of rho or N should be specified.

alpha

Numeric. The desired alpha level. Default: 0.05

alternative

Character. Specifies the alternative hypothesis. Must be: two_sided (Default), less, or greater

power

Numeric. The desired power. Must be a value between 0 and 1. Default: 0.8.

sensitivity

Numeric. The sensitivity of the linkage criteria. Must be between 0 and 1. Default: 1.

specificity

Numeric. The specificity of the linkage criteria. Must be between 0 and 1. Default: 1.

overdispersion

Numeric (Positive). An overdispersion parameter, set if the assumed distribution of the number of edges is negative binomial. If NULL the assumed distribution is Poisson (equivalent to an overdispersion parameter of infinity) Default: NULL Note that this is equivalent to setting the overdispersion parameter to Inf.

Value

Squared error between the input sample size and estimated sample size