Calculate sample size

This function calculates the sample size needed to obtain at least a defined false disovery rate given a final outbreak size \(N\).

samplesize(eta, chi, N, R = NULL, phi, min_pairs = 1, assumption = "mtml")

Arguments

eta
scalar or vector giving the sensitivity of the linkage criteria

chi
scalar or vector giving the specificity of the linkage criteria

N
scalar or vector giving the final outbreak size

R
scalar or vector giving the effective reproductive number of the pathogen

phi
scalar or vector giving the desired true discovery rate (1-false discovery rate)

min_pairs
minimum number of linked pairs observed in the sample, defaults to 1 pair (2 samples); this is to ensure reasonable results are obtained

assumption
a character vector indicating which assumptions about transmission and linkage criteria. The function expects one of the following:

'stsl' for the single-transmission single-linkage assumption (prob_trans_stsl).

'mtsl' for the multiple-transmission single-linkage assumption (prob_trans_mtsl).

'mtml' for the multiple-transmission multiple-linkage assumption (prob_trans_mtml).

Value

scalar or vector giving the sample size needed to meet the given conditions

Examples

samplesize(eta=0.99, chi=0.995, N=100, R=1, phi=0.75)

#> [1] 10