Expected number of true transmission pairs assuming multiple-transmission and single-linkage

This function calculates the expected number true transmission pairs in a sample of size M. The multiple-transmission and single-linkage method assumes the following:

Each case \(i\) is, on average, the infector of R cases in the population (\(N\))

Each case \(i\) is allowed to be linked by the linkage criteria to only one other case \(j\) in the sampled population (\(M\)).

true_pairs_mtsl(eta, rho, M, R)

Arguments

eta

scalar or vector giving the sensitivity of the linkage criteria

rho

scalar or vector giving the proportion of the final outbreak size that is sampled

M

scalar or vector giving the number of cases sampled

R

scalar or vector giving the effective reproductive number of the pathogen

Value

scalar or vector giving the expected number of true transmission pairs in the sample

See also

Author

John Giles, Shirlee Wohl and Justin Lessler

Examples

true_pairs_mtsl(eta=0.95, rho=0.2, M=200, R=1)
#> Warning: `true_pairs_mtsl()` was deprecated in phylosamp 1.0.0.
#> ℹ Please use `translink_expected_links_true_mtsl()` instead.
#> Warning: The `eta` argument of `true_pairs_mtsl()` is deprecated as of phylosamp 1.0.0.
#> ℹ Please use the `sensitivity` argument of
#>   `translink_expected_links_true_mtsl()` instead.
#> [1] 38